catching up after a while: renaming, comments
