+++ setup/constants.php
@@ -3,5 +3,5 @@
 namespace Launchpad\Setup;
 
 const TITLE = 'Launchpad';
-const NAME = 'launchpad'; // Constants can't be variables
+const NAME = 'launchpad';
 const VERSION = '2.0.0';
